Discophone

+++ egs/discophone/v1/run.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
+# Copyright 2020  Johns Hopkins University (Author: Piotr Żelasko)
+# Apache 2.0 (http://www.apache.org/licenses/LICENSE-2.0)
 
 set -eou pipefail
 
 stage=0
 train_nj=24
 phone_tokens=false
+phone_ngram_order=2
 
 # Acoustic model parameters
 numLeavesTri1=1000
@@ -40,16 +43,13 @@ else
   mboshi_train=false
   mboshi_recog=false
   gp_romanized=false
-  ipa_transcript=true
 fi
 
 . cmd.sh
 . utils/parse_options.sh
 . path.sh
 
-
 local/install_shorten.sh
-
 
 train_set=""
 dev_set=""
@@ -82,9 +82,20 @@ function langname() {
   echo "$(basename "$1")"
 }
 
+phone_token_opt='--phones'
+if [ $phone_tokens = true ]; then
+  phone_token_opt='--phone-tokens'
+fi
+
+# This step will create the data directories for GlobalPhone and Babel languages.
+# It's also going to use LanguageNet G2P models to convert text into phonetic transcripts.
+# Depending on the settings, it will either transcribe into phones, e.g. ([m], [i:], [t]), or
+# phonetic tokens, e.g. (/m/, /i/, /:/, /t/).
+# The Kaldi "text" file will consist of these phonetic sequences, as we're trying to build
+# a universal IPA recognizer.
+# The lexicons are created separately for each split as an artifact from the ESPnet setup.
 if ((stage <= 0)); then
   echo "stage 0: Setting up individual languages"
-
   local/setup_languages.sh \
     --langs "${babel_langs}" \
     --recog "${babel_recog}" \
@@ -93,36 +104,53 @@ if ((stage <= 0)); then
     --mboshi-train "${mboshi_train}" \
     --mboshi-recog "${mboshi_recog}" \
     --gp-romanized "${gp_romanized}" \
-    --ipa-transcript "${ipa_transcript}"
+    --gp-path "${gp_path}" \
+    --phone_token_opt "${phone_token_opt}"
   for x in ${train_set} ${dev_set} ${recog_set}; do
     sed -i.bak -e "s/$/ sox -R -t wav - -t wav - rate 16000 dither | /" data/${x}/wav.scp
   done
-
-  for data_dir in ${train_set}; do
-    if [ -f data/$data_dir/text.bkp ]; then
-      # replace IPA text with normal text
-      cp data/$data_dir/text.bkp data/$data_dir/text
-    fi
-  done
 fi
 
-phone_token_opt=
-if [ $phone_tokens = true ]; then
-  phone_token_opt='--phone-tokens'
-fi
-
-if ((stage <= 4)); then
+# Here we will combine the lexicons for train/dev/test splits into a single lexicon for each language.
+if ((stage <= 2)); then
   for data_dir in ${train_set}; do
     lang_name=$(langname $data_dir)
     mkdir -p data/local/$lang_name
-    python local/prepare_lexicon_dir.py $phone_token_opt data/$data_dir/lexicon_ipa.txt data/local/$lang_name
-    lang_name="$(langname $data_dir)"
-    utils/prepare_lang.sh \
-      --share-silence-phones true \
-      data/local/$lang_name '<unk>' data/local/tmp.lang/$lang_name data/lang/$lang_name
+    python3 local/combine_lexicons.py \
+      data/$data_dir/lexicon_ipa.txt \
+      data/${data_dir//train/dev}/lexicon_ipa.txt \
+      data/${data_dir//train/eval}/lexicon_ipa.txt \
+      >data/$data_dir/lexicon_ipa_all.txt
+    python3 local/prepare_lexicon_dir.py $phone_token_opt data/$data_dir/lexicon_ipa_all.txt data/local/$lang_name
   done
 fi
 
+# We use the per-language lexicons to find the set of phones/phonetic tokens in every language and combine
+# them again to obtain a multilingual "dummy" lexicon of the form:
+# a a
+# b b
+# c c
+# ...
+# When that is ready, we train a multilingual phone-level language model (i.e. phonotactic model),
+# that will be used to compile the decoding graph and to score each ASR system.
+if ((stage <= 3)); then
+  local/prepare_ipa_lm.sh \
+    --train-set "$train_set" \
+    --phone_token_opt "$phone_token_opt" \
+    --order "$phone_ngram_order"
+  lexicon_list=$(find data/ipa_lm/train -name lexiconp.txt)
+  mkdir -p data/local/dict_combined/local
+  python3 local/combine_lexicons.py $lexicon_list >data/local/dict_combined/local/lexiconp.txt
+  python3 local/prepare_lexicon_dir.py data/local/dict_combined/local/lexiconp.txt data/local/dict_combined
+  utils/prepare_lang.sh \
+    --position-dependent-phones false \
+    data/local/dict_combined \
+    "<unk>" data/local/dict_combined data/lang_combined
+  LM=data/ipa_lm/train_all/srilm.o${phone_ngram_order}g.kn.gz
+  utils/format_lm.sh data/lang_combined "$LM" data/local/dict_combined/lexicon.txt data/lang_combined_test
+fi
+
+# MFCC extraction for GMM training
 if ((stage <= 5)); then
   # Feature extraction
   for data_dir in ${train_set}; do
@@ -130,10 +158,10 @@ if ((stage <= 5)); then
       lang_name=$(langname $data_dir)
       steps/make_mfcc.sh \
         --cmd "$train_cmd" \
-        --nj 8 \
+        --nj 16 \
         --write_utt2num_frames true \
-        data/$data_dir \
-        exp/make_mfcc/$data_dir \
+        "data/$data_dir" \
+        "exp/make_mfcc/$data_dir" \
         mfcc
       utils/fix_data_dir.sh data/$data_dir
       steps/compute_cmvn_stats.sh data/$data_dir exp/make_mfcc/$lang_name mfcc/$lang_name
@@ -151,19 +179,25 @@ if ((stage <= 6)); then
       utils/subset_data_dir.sh data/$data_dir 5000 data/subsets/5k/$data_dir
     else
       mkdir -p "$(dirname data/subsets/5k/$data_dir)"
-      ln -s "$(pwd)/data/$data_dir" "data/subsets/5k/$data_dir"
+      if [ ! -L "data/subsets/5k/$data_dir" ]; then
+        ln -s "$(pwd)/data/$data_dir" "data/subsets/5k/$data_dir"
+      fi
     fi
     if [ $numutt -gt 10000 ]; then
       utils/subset_data_dir.sh data/$data_dir 10000 data/subsets/10k/$data_dir
     else
       mkdir -p "$(dirname data/subsets/10k/$data_dir)"
-      ln -s "$(pwd)/data/$data_dir" "data/subsets/10k/$data_dir"
+      if [ ! -L "data/subsets/10k/$data_dir" ]; then
+        ln -s "$(pwd)/data/$data_dir" "data/subsets/10k/$data_dir"
+      fi
     fi
     if [ $numutt -gt 20000 ]; then
       utils/subset_data_dir.sh data/$data_dir 20000 data/subsets/20k/$data_dir
     else
       mkdir -p "$(dirname data/subsets/20k/$data_dir)"
-      ln -s "$(pwd)/data/$data_dir" "data/subsets/20k/$data_dir"
+      if [ ! -L "data/subsets/20k/$data_dir" ]; then
+        ln -s "$(pwd)/data/$data_dir" "data/subsets/20k/$data_dir"
+      fi
     fi
   done
 fi
@@ -177,7 +211,7 @@ if ((stage <= 7)); then
       steps/train_mono.sh \
         --nj 8 --cmd "$train_cmd" \
         data/subsets/5k/$data_dir \
-        data/lang/$lang_name $expdir
+        data/lang_combined_test $expdir
     ) &
     sleep 2
   done
@@ -192,7 +226,7 @@ if ((stage <= 8)); then
       steps/align_si.sh \
         --nj 12 --cmd "$train_cmd" \
         data/subsets/10k/$data_dir \
-        data/lang/$lang_name \
+        data/lang_combined_test \
         exp/gmm/$lang_name/mono \
         exp/gmm/$lang_name/mono_ali_10k
 
@@ -201,7 +235,7 @@ if ((stage <= 8)); then
         $numLeavesTri1 \
         $numGaussTri1 \
         data/subsets/10k/$data_dir \
-        data/lang/$lang_name \
+        data/lang_combined_test \
         exp/gmm/$lang_name/mono_ali_10k \
         exp/gmm/$lang_name/tri1
     ) &
@@ -218,25 +252,16 @@ if ((stage <= 9)); then
       steps/align_si.sh \
         --nj 24 --cmd "$train_cmd" \
         data/subsets/20k/$data_dir \
-        data/lang/$lang_name \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri1 \
         exp/gmm/$lang_name/tri1_ali_20k
 
       steps/train_deltas.sh \
         --cmd "$train_cmd" $numLeavesTri2 $numGaussTri2 \
         data/subsets/20k/$data_dir \
-        data/lang/$lang_name \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri1_ali_20k \
         exp/gmm/$lang_name/tri2
-
-      local/reestimate_langp.sh --cmd "$train_cmd" --unk "<unk>" \
-        data/subsets/20k/$data_dir \
-        data/lang/$lang_name \
-        data/local/$lang_name \
-        exp/gmm/$lang_name/tri2 \
-        data/local/dictp/$lang_name/tri2 \
-        data/local/langp/$lang_name/tri2 \
-        data/langp/$lang_name/tri2
     ) &
     sleep 2
   done
@@ -251,25 +276,16 @@ if ((stage <= 10)); then
       steps/align_si.sh \
         --nj $train_nj --cmd "$train_cmd" \
         data/$data_dir \
-        data/langp/$lang_name/tri2 \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri2 \
         exp/gmm/$lang_name/tri2_ali
 
       steps/train_deltas.sh \
         --cmd "$train_cmd" $numLeavesTri3 $numGaussTri3 \
         data/$data_dir \
-        data/langp/$lang_name/tri2 \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri2_ali \
         exp/gmm/$lang_name/tri3
-
-      local/reestimate_langp.sh --cmd "$train_cmd" --unk "<unk>" \
-        data/$data_dir \
-        data/lang/$lang_name \
-        data/local/$lang_name \
-        exp/gmm/$lang_name/tri3 \
-        data/local/dictp/$lang_name/tri3 \
-        data/local/langp/$lang_name/tri3 \
-        data/langp/$lang_name/tri3
     ) &
     sleep 2
   done
@@ -284,7 +300,7 @@ if ((stage <= 11)); then
       steps/align_si.sh \
         --nj $train_nj --cmd "$train_cmd" \
         data/$data_dir \
-        data/langp/$lang_name/tri3 \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri3 \
         exp/gmm/$lang_name/tri3_ali
 
@@ -293,18 +309,9 @@ if ((stage <= 11)); then
         $numLeavesMLLT \
         $numGaussMLLT \
         data/$data_dir \
-        data/langp/$lang_name/tri3 \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri3_ali \
         exp/gmm/$lang_name/tri4
-
-      local/reestimate_langp.sh --cmd "$train_cmd" --unk "<unk>" \
-        data/$data_dir \
-        data/lang/$lang_name \
-        data/local/$lang_name \
-        exp/gmm/$lang_name/tri4 \
-        data/local/dictp/$lang_name/tri4 \
-        data/local/langp/$lang_name/tri4 \
-        data/langp/$lang_name/tri4
     ) &
     sleep 2
   done
@@ -319,7 +326,7 @@ if ((stage <= 12)); then
       steps/align_si.sh \
         --nj $train_nj --cmd "$train_cmd" \
         data/$data_dir \
-        data/langp/$lang_name/tri4 \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri4 \
         exp/gmm/$lang_name/tri4_ali
 
@@ -328,18 +335,9 @@ if ((stage <= 12)); then
         $numLeavesSAT \
         $numGaussSAT \
         data/$data_dir \
-        data/langp/$lang_name/tri4 \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri4_ali \
         exp/gmm/$lang_name/tri5
-
-      local/reestimate_langp.sh --cmd "$train_cmd" --unk "<unk>" \
-        data/$data_dir \
-        data/lang/$lang_name \
-        data/local/$lang_name \
-        exp/gmm/$lang_name/tri5 \
-        data/local/dictp/$lang_name/tri5 \
-        data/local/langp/$lang_name/tri5 \
-        data/langp/$lang_name/tri5
     ) &
     sleep 2
   done
@@ -354,18 +352,9 @@ if ((stage <= 13)); then
       steps/align_fmllr.sh \
         --nj $train_nj --cmd "$train_cmd" \
         data/$data_dir \
-        data/langp/$lang_name/tri5 \
+        data/lang_combined_test \
         exp/gmm/$lang_name/tri5 \
         exp/gmm/$lang_name/tri5_ali
-
-      local/reestimate_langp.sh --cmd "$train_cmd" --unk "<unk>" \
-        data/$data_dir \
-        data/lang/$lang_name \
-        data/local/$lang_name \
-        exp/gmm/$lang_name/tri5_ali \
-        data/local/dictp/$lang_name/tri5_ali \
-        data/local/langp/$lang_name/tri5_ali \
-        data/langp/$lang_name/tri5_ali
     ) &
     sleep 2
   done
